+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Remove "layer" from service directive for now,
+  because it is not yet implemented properly.
+  [jone]
 
 
 1.0a1 (2015-08-01)